Move utils to root Home folder
